% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzz_articles.R
\name{clean_nzz_articles}
\alias{clean_nzz_articles}
\title{Cleaning raw NZZ articles}
\usage{
clean_nzz_articles(nzz_articles, no_updates = TRUE)
}
\arguments{
\item{nzz_articles}{a data frame containing raw NZZ articles as returned from the function
\code{get_nzz_articles}.}

\item{no_updates}{Boolean that indicates if updated versions should be removed.
If TRUE, each article kept is unique (default), else also updated versions of articles are kept.}
}
\value{
returns the cleaned data frame with the same columns.
}
\description{
Cleaning raw NZZ articles
}
\examples{

# not run
# nzz_articles_raw <- get_nzz_articles()
# nzz_articles_clean <- clean_nzz_articles(nzz_articles_raw)

}
