% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzz_articles.R
\name{clean_nzz_authors}
\alias{clean_nzz_authors}
\title{Helper function to clean NZZ articles}
\usage{
clean_nzz_authors(nzz_articles)
}
\arguments{
\item{nzz_articles}{a data frame containing raw NZZ articles as returned from the function
\code{get_nzz_articles}.}
}
\value{
the cleaned data frame \code{nzz_articles}.
}
\description{
Removes redundant information from NZZ author column and separates collaborating authors into two
separate author columns.
}
