% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzz_articles.R
\name{clean_nzz_dates}
\alias{clean_nzz_dates}
\title{A helper function to clean NZZ articles}
\usage{
clean_nzz_dates(nzz_articles)
}
\arguments{
\item{nzz_articles}{a data frame containing raw NZZ articles as returned from the function
\code{get_nzz_articles}.}
}
\value{
the cleaned data frame \code{nzz_articles}
}
\description{
Turns NZZ published_time into a day, month and year date column, in which the article was published.
}
