% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srf_articles.R
\name{clean_srf_articles}
\alias{clean_srf_articles}
\title{Cleaning raw SRF articles}
\usage{
clean_srf_articles(srf_articles, no_updates = TRUE)
}
\arguments{
\item{srf_articles}{a data frame containing raw SRF articles as returned from the functions
\code{get_srf_articles} or \code{get_srf_articles_continously}.}

\item{no_updates}{Boolean that indicates if updated versions should be removed.
If TRUE, each article kept is unique (default), else also updated versions of articles are kept.}
}
\value{
returns the cleaned data frame with the same columns.
}
\description{
Cleaning raw SRF articles
}
\examples{

# not run
# srf_articles_raw <- get_srf_articles()
# srf_articles_clean <- clean_srf_articles(srf_articles_raw)

}
