% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srf_articles.R
\name{clean_srf_authors}
\alias{clean_srf_authors}
\title{Helper function to clean SRF articles}
\usage{
clean_srf_authors(srf_articles)
}
\arguments{
\item{srf_articles}{a data frame containing raw SRF articles as returned from the functions
\code{get_srf_articles} or \code{get_srf_articles_continously}.}
}
\value{
the cleaned data frame \code{srf_articles}.
}
\description{
Removes redundant information from SRF author column and separates collaborating authors into two
separate author columns.
}
