% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srf_articles.R
\name{clean_srf_dates}
\alias{clean_srf_dates}
\title{A helper function to clean SRF articles}
\usage{
clean_srf_dates(srf_articles)
}
\arguments{
\item{srf_articles}{a data frame containing raw SRF articles as returned from the functions
\code{get_srf_articles} or \code{get_srf_articles_continously}.}
}
\value{
the cleaned data frame \code{srf_articles}
}
\description{
Turns SRF uploadDate into a day, month and year date column, in which the article was published.
}
