% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woz_articles.R
\name{clean_woz_articles}
\alias{clean_woz_articles}
\title{Cleaning raw WOZ articles}
\usage{
clean_woz_articles(woz_articles, no_updates = TRUE)
}
\arguments{
\item{woz_articles}{a data frame containing raw WOZ articles as returned from the functions
\code{get_woz_articles}, \code{get_woz_articles_continously},
or \code{get_woz_articles_from_archive}.}

\item{no_updates}{Boolean that indicates if updated versions should be removed.
If TRUE, each article kept is unique (default), else also updated versions of articles are kept.}
}
\value{
returns the cleaned data frame with the same columns.
}
\description{
Cleaning raw WOZ articles
}
\examples{

# not run
# woz_articles_raw <- get_woz_articles()
# woz_articles_clean <- clean_woz_articles(woz_articles_raw)

}
