% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woz_articles.R
\name{clean_woz_dates}
\alias{clean_woz_dates}
\title{A helper function to clean WOZ articles}
\usage{
clean_woz_dates(woz_articles)
}
\arguments{
\item{woz_articles}{a data frame containing raw WOZ articles as returned from the functions
\code{get_woz_articles}, \code{get_woz_articles_continously},
or \code{get_woz_articles_from_archive}}
}
\value{
the cleaned data frame \code{woz_articles}
}
\description{
Turns WOZ field issue into a day, month and year date column, in which the article was published.
}
