% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzz_articles.R
\name{construct_nzz_df_skeleton}
\alias{construct_nzz_df_skeleton}
\title{Constructs a data frame with the required column names for the function \code{get_nzz_articles()}}
\usage{
construct_nzz_df_skeleton()
}
\value{
a data frame with 13 named empty columns (see details).
}
\description{
Constructs a data frame with the required column names for the function \code{get_nzz_articles()}
}
\details{
The following empty columns are contained in the returned df:
\itemize{
 \item{"full_link"}
 \item{"section"}
 \item{"description"}
 \item{"author}
 \item{"published_time"}
 \item{"modified_time"}
 \item{"title"}
 \item{"headline_title"}
 \item{"headline_lead}
 \item{"image_caption}
 \item{"image_author"}
 \item{"text"}
 \item{"linked_authors"}
}
}
\examples{
nzz_articles_empty <- construct_nzz_df_skeleton()

}
