% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srf_articles.R
\name{construct_srf_df_skeleton}
\alias{construct_srf_df_skeleton}
\title{Constructs a data frame with the required column names for the functions \code{get_srf_articles()}}
\usage{
construct_srf_df_skeleton()
}
\value{
a data frame with 27 named empty columns (see details).
}
\description{
Constructs a data frame with the required column names for the functions \code{get_srf_articles()}
}
\details{
The following empty columns are contained in the returned df:
\itemize{
 \item{"uploadDate"}
 \item{"description"}
 \item{"thumbnailUrl"}
 \item{"section"}
 \item{"published_time"}
 \item{"modified_time"}
 \item{"identifier}
 \item{"image_url}
 \item{"commentCount}
 \item{"full_link}
 \item{"section_link"}
 \item{"subsection"}
 \item{"title_overline"}
 \item{"title"}
 \item{"subtitle"}
 \item{"author}
 \item{"sharings}
 \item{"text_titles}
 \item{"text"}
 \item{"quote"}
 \item{"expandable_box"}
 \item{"related_articles"}
 \item{"related_articles_info"}
 \item{"reference}
 \item{"allows_comments}
 \item{"comment_author}
 \item{"comment_content"}
}
}
\examples{
srf_articles_empty <- construct_srf_df_skeleton()

}
