% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woz_articles.R
\name{construct_woz_df_skeleton}
\alias{construct_woz_df_skeleton}
\title{Constructs a data frame with the required column names for the functions \code{get_woz_articles()}}
\usage{
construct_woz_df_skeleton()
}
\value{
a data frame with 13 named empty columns (see details).
}
\description{
Constructs a data frame with the required column names for the functions \code{get_woz_articles()}
}
\details{
The following empty columns are contained in the returned df:
\itemize{
 \item{"section"}
 \item{"published_time"}
 \item{"modified_time"}
 \item{"full_link}
 \item{"field_issue"}
 \item{"title"}
 \item{"subtitle"}
 \item{"author}
 \item{"image_link}
 \item{"image_caption}
 \item{"text_titles}
 \item{"text"}
 \item{"prowoz"}
}
}
\examples{
woz_articles_empty <- construct_woz_df_skeleton()

}
