% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srf_articles.R
\name{get_article_data}
\alias{get_article_data}
\title{Helper Function to extract data of raw srf html-document and return it as vector.}
\usage{
get_article_data(full_link)
}
\arguments{
\item{full_link}{string with link to srf article}
}
\description{
Helper Function to extract data of raw srf html-document and return it as vector.
}
