% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzz_articles.R
\name{get_newest_nzz_article_links}
\alias{get_newest_nzz_article_links}
\title{Function to fetch newest NZZ article links}
\usage{
get_newest_nzz_article_links(n = 1000)
}
\arguments{
\item{n}{integer indicating the number of most recent articles to download. The maximum possible is 10000.}
}
\value{
vector of length \code{n} containing the NZZ article links as characters.
}
\description{
Function to fetch newest NZZ article links
}
