% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srf_articles.R
\name{get_newest_srf_articles}
\alias{get_newest_srf_articles}
\title{Helper Function to get newest srf article links and extract article content}
\usage{
get_newest_srf_articles(srf_articles)
}
\arguments{
\item{srf_articles}{a data frame containing previously fetched articles (can be zero)}
}
\description{
Helper Function to get newest srf article links and extract article content
}
