% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzz_articles.R
\name{get_nzz_article_data}
\alias{get_nzz_article_data}
\title{Helper Function to extract article data of raw nzz html-document and return it as vector.}
\usage{
get_nzz_article_data(raw_article)
}
\arguments{
\item{raw_article}{list element with one raw nzz article as returned from \code{get_nzz_articles()}.}
}
\description{
Helper Function to extract article data of raw nzz html-document and return it as vector.
}
