% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzz_articles.R
\name{get_nzz_articles_raw}
\alias{get_nzz_articles_raw}
\title{Function to get raw nzz articles}
\usage{
get_nzz_articles_raw(article_links)
}
\arguments{
\item{article_links}{string vector of length 1 or more wi links to the nzz articles}
}
\value{
list containing the raw nzz articles
}
\description{
The nzz articles according to the provided link are downloaded and stored in a list.
}
\examples{

## not run
# links <- get_newest_nzz_article_links()
# get_nzz_articles_raw(links)

}
