% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woz_articles.R
\name{get_section_articles}
\alias{get_section_articles}
\title{Get the newest articles from each section on woz.ch}
\usage{
get_section_articles(woz_articles, section)
}
\arguments{
\item{woz_articles}{a data frame with 13 named columns (see \code{get_woz_articles} and \code{construct_woz_df_skeleton} )}

\item{section}{sections to scrape on woz.ch. Default is all the permanent sections currently (2022-04-06) on the website.}
}
\value{
the data frame \code{woz_articles} updated with all the fetched articles.
}
\description{
Get the newest articles from each section on woz.ch
}
