% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srf_articles.R
\name{get_srf_articles}
\alias{get_srf_articles}
\title{Get SRF articles once when executed}
\usage{
get_srf_articles(srf_articles = NULL)
}
\arguments{
\item{srf_articles}{either NULL (default) or a data frame containing previously fetched articles (see details).
IF NULL, an empty skeleton is automatically provided with \code{construct_srf_df_skeleton()}.
Else, the new or updated articles are appended to exisiting data frame \code{srf_articles}.}
}
\value{
the data frame \code{srf_articles} updated with all the fetched articles. Also stores the df as csv in the background.
}
\description{
Function fetches the most recent online articles of the Swiss news institution "Schweizer Radio und Fernsehen (SRF)".
The function can run daily, or even less frequently. Especially useful if executed with cronjob or manually on a daily basis.
}
\details{
The following columns need to be contained in the df \code{srf_articles} if it is provided as parameter:
\code{srf_articles}:
\itemize{
 \item{"uploadDate"}
 \item{"description"}
 \item{"thumbnailUrl"}
 \item{"section"}
 \item{"published_time"}
 \item{"modified_time"}
 \item{"identifier}
 \item{"image_url}
 \item{"commentCount}
 \item{"full_link}
 \item{"section_link"}
 \item{"subsection"}
 \item{"title_overline"}
 \item{"title"}
 \item{"subtitle"}
 \item{"author}
 \item{"sharings}
 \item{"text_titles}
 \item{"text"}
 \item{"quote"}
 \item{"expandable_box"}
 \item{"related_articles"}
 \item{"related_articles_info"}
 \item{"reference}
 \item{"allows_comments}
 \item{"comment_author}
 \item{"comment_content"}
}
}
\examples{
# not run
# get_srf_articles()

}
