% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srf_articles.R
\name{get_srf_articles_continously}
\alias{get_srf_articles_continously}
\title{Get newest SRF articles over a defined period of time}
\usage{
get_srf_articles_continously(
  srf_articles = NULL,
  update_in_minutes = 15,
  total_updates = 48,
  save_backup = TRUE
)
}
\arguments{
\item{srf_articles}{either NULL (default) or a data frame containing previously fetched articles (see details).
IF NULL, an empty skeleton is automatically provided with \code{construct_srf_df_skeleton()}.
Else, the new or updated articles are appended to exisiting data frame \code{srf_articles}.}

\item{update_in_minutes}{an integer that defines after how many minutes the function should check for new or updated articles.}

\item{total_updates}{an integer that defines the total amount of updates. Multiplying \code{total_updates} with \code{update_in_minutes} gives a rough approximation of total execution time.}

\item{save_backup}{If TRUE (default) stores the articles as csv-file after every mutation as backup.}
}
\value{
the data frame \code{srf_articles} updated with all the fetched articles.
}
\description{
Function regularly fetches the most recent online articles of the Swiss news institution "Schweizer Radio und Fernsehen (SRF)".
This function is especially useful to analyze if and how articles are being updated after they have been published online.
}
\details{
The following columns need to exist in the df \code{srf_articles}, if it is provided as parameter:
\code{srf_articles}:
\itemize{
 \item{"uploadDate"}
 \item{"description"}
 \item{"thumbnailUrl"}
 \item{"section"}
 \item{"published_time"}
 \item{"modified_time"}
 \item{"identifier}
 \item{"image_url}
 \item{"commentCount}
 \item{"full_link}
 \item{"section_link"}
 \item{"subsection"}
 \item{"title_overline"}
 \item{"title"}
 \item{"subtitle"}
 \item{"author}
 \item{"sharings}
 \item{"text_titles}
 \item{"text"}
 \item{"quote"}
 \item{"expandable_box"}
 \item{"related_articles"}
 \item{"related_articles_info"}
 \item{"reference}
 \item{"allows_comments}
 \item{"comment_author}
 \item{"comment_content"}
}
}
\examples{
# not run
# get_srf_articles_continously(srf_articles = NULL, update_in_minutes = 5, total_updates = 50)

}
