% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woz_articles.R
\name{get_woz_article_data}
\alias{get_woz_article_data}
\title{Helper Function to extract data of raw woz html-document and return it as vector.}
\usage{
get_woz_article_data(full_link, section = NULL)
}
\arguments{
\item{full_link}{string with link to woz article}

\item{section}{sections to scrape on woz.ch. Default is all the permanent sections currently (2022-04-06) on the website.}
}
\description{
Helper Function to extract data of raw woz html-document and return it as vector.
}
