% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woz_articles.R
\name{get_woz_articles}
\alias{get_woz_articles}
\title{Get WOZ articles once when executed}
\usage{
get_woz_articles(
  woz_articles = NULL,
  sections = c("schweiz", "wirtschaft", "international", "kultur-wissen")
)
}
\arguments{
\item{woz_articles}{either NULL (default) or a data frame containing previously fetched articles (see details).
IF NULL, an empty skeleton is automatically provided with \code{construct_woz_df_skeleton()}.
Else, the new or updated articles are appended to exisiting data frame \code{woz_articles}.}

\item{sections}{sections to scrape on woz.ch. Default is all the permanent sections currently (2022-04-06) on the website.}
}
\value{
the data frame \code{woz_articles} updated with all the fetched articles. Also stores the df as csv in the background.
}
\description{
Function fetches the most recent online articles of the Swiss news outlet "Die Wochenzeitung (WOZ)".
The function can be executed daily or less frequently. It is especially useful if executed with cronjob or manually on a daily basis.
}
\details{
The following columns need to be contained in the df \code{woz_articles} if it is provided as parameter:
\code{woz_articles}:
\itemize{
 \item{"section"}
 \item{"published_time"}
 \item{"modified_time"}
 \item{"full_link}
 \item{"field_issue"}
 \item{"title"}
 \item{"subtitle"}
 \item{"author}
 \item{"image_link}
 \item{"image_caption}
 \item{"text_titles}
 \item{"text"}
 \item{"prowoz"}
}
}
\examples{
# not run
# get_woz_articles()

}
