% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woz_articles.R
\name{get_woz_meta_data}
\alias{get_woz_meta_data}
\title{Helper Function to extract meta data of raw woz html-document and return it as vector.}
\usage{
get_woz_meta_data(full_link)
}
\arguments{
\item{full_link}{string with link to woz article}
}
\description{
Helper Function to extract meta data of raw woz html-document and return it as vector.
}
