% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{plot_author_frequency}
\alias{plot_author_frequency}
\title{Plotting Article per Author Frequency}
\usage{
plot_author_frequency(
  articles,
  period = "all",
  top = 5,
  specific_author = NULL,
  from = NULL,
  to = NULL
)
}
\arguments{
\item{articles}{data frame with articles to be counted and plotted. Needs to contain a date (yyyy-mm-dd) column
\code{month_published}, and optionally \code{day_published} or \code{year_published},
if articles should be plotted per day or year. Also needs columns \code{author_1} and \code{author_2}
to count the articles per author.}

\item{period}{string indicating the time period over which the articles should be published. Either articles are counted per
overall indicated with "all" (default) or per "day", "month", or "year".}

\item{top}{integer indicating which authors should be plotted. If 5 (default), the five authors with overall
the most number of articles are plotted.}

\item{specific_author}{vector string of length 1 or more, containing the authors to be plotted. If NULL (default),
the five (or as indicated in "top") most frequent authors are being plotted.}

\item{from}{string i.e. date of format "yyyy-mm-dd" indicating the publishing date from which on articles should be evaluated and plotted.
For the time period to be evaluated both `from` and `to` need to contain a date. The time period must include at least one first of January.}

\item{to}{string i.e. date of format "yyyy-mm-dd" indicating the publishing date until which articles should be evaluated and plotted.
For the time period to be evaluated both `from` and `to` need to contain a date. The time period must include at least one first of January.}
}
\value{
a ggplot bar or line chart with the article frequency per author
}
\description{
Plots the number of articles published by each author either in total in a bar chart or
per day, month, or year in a line chart.
}
\examples{

## not run
# woz_articles_raw <- get_woz_articles()
# woz_articles_clean <- clean_woz_articles(woz_articles_raw)
# plot_article_frequency(woz_articles_clean)

}
