% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{plot_section_frequency}
\alias{plot_section_frequency}
\title{Plotting Article per Section Frequency}
\usage{
plot_section_frequency(
  articles,
  period = "all",
  top = 5,
  specific_section = NULL,
  relative = FALSE,
  from = NULL,
  to = NULL
)
}
\arguments{
\item{articles}{data frame with articles to be counted and plotted. Needs to contain a date (yyyy-mm-dd) column
\code{month_published}, and optionally \code{day_published} or \code{year_published},
if articles should be plotted per day or year.}

\item{period}{string indicating the time period over which the articles should be published. Either articles are counted per
overall indicated with "all" (default) or per "day", "month", or "year".}

\item{top}{integer indicating the which sections should be plotted. If 5 (default), the five sections with overall
the most number of articles are plotted.}

\item{specific_section}{vector string of length 1 or more, containing the sections to be plotted. If NULL (default),
the five (or as indicated in "top") most frequent sections are being plotted.}

\item{relative}{IF FALSE (default) the absolute number of articles per section are plotted.
If TRUE, they are plotted relatively to all articles in the selected sections.}

\item{from}{string i.e. date of format "yyyy-mm-dd" indicating the publishing date from which on articles should be evaluated and plotted.
For the time period to be evaluated both `from` and `to` need to contain a date. The time period must include at least one first of January.}

\item{to}{string i.e. date of format "yyyy-mm-dd" indicating the publishing date until which articles should be evaluated and plotted.
For the time period to be evaluated both `from` and `to` need to contain a date. The time period must include at least one first of January.}
}
\value{
a ggplot bar or line chart with the article frequency per section
}
\description{
Plots the number of articles published in each section (german "Ressort") either in total in a bar chart or
per day, month, or year in a line chart.
}
\examples{

## not run
# woz_articles_raw <- get_woz_articles()
# woz_articles_clean <- clean_woz_articles(woz_articles_raw)
# plot_article_frequency(woz_articles_clean)

}
