% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{plot_topic_frequency}
\alias{plot_topic_frequency}
\title{Plotting Frequency of Topics in Articles}
\usage{
plot_topic_frequency(
  articles,
  topic,
  period = "all",
  distinct_only = FALSE,
  show_total = FALSE,
  relative = FALSE,
  from = NULL,
  to = NULL
)
}
\arguments{
\item{articles}{data frame with articles to be counted and plotted. Needs to contain a date (yyyy-mm-dd) column
\code{month_published}, and optionally \code{day_published} or \code{year_published},
if articles should be plotted per day or year.}

\item{topic}{string vector of length 1 or more with the words to search in the article texts.}

\item{period}{string indicating the time period over which the articles should be published. Either articles are counted per
overall indicated with "all" (default) or per "day", "month", or "year".}

\item{distinct_only}{If FALSE (default), also plots articles where a combination of words in \code{topic} is present.
If TRUE, plots only those articles where only one of the words is present.}

\item{show_total}{If FALSE (default), articles are plotted by topic. If TRUE all articles are plotted that contain at least one word in \code{topic}.}

\item{relative}{IF FALSE (default) the absolute number of articles per section are plotted.
If TRUE, they are plotted relatively to all articles in the selected sections.}

\item{from}{string i.e. date of format "yyyy-mm-dd" indicating the publishing date from which on articles should be evaluated and plotted.
For the time period to be evaluated both `from` and `to` need to contain a date. The time period must include at least one first of January.}

\item{to}{string i.e. date of format "yyyy-mm-dd" indicating the publishing date until which articles should be evaluated and plotted.
For the time period to be evaluated both `from` and `to` need to contain a date. The time period must include at least one first of January.}
}
\value{
a ggplot bar or line chart with the article frequency per section
}
\description{
Plotting Frequency of Topics in Articles
}
\examples{

## not run
# woz_articles_raw <- get_woz_articles()
# woz_articles_clean <- clean_woz_articles(woz_articles_raw)
# topic <- c("Umweltschutz", "Altersvorsorge", "Krankenkassen")
# plot_topic_frequency(woz_articles_clean, topic)

}
