% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzz_articles.R
\name{separate_nzz_authors}
\alias{separate_nzz_authors}
\title{A helper function to clean NZZ authors}
\usage{
separate_nzz_authors(nzz_articles)
}
\arguments{
\item{nzz_articles}{a data frame containing raw NZZ articles as returned from the function
\code{get_nzz_articles}.}
}
\value{
the cleaned data frame \code{nzz_articles}
}
\description{
Separates collaborating authors into two separate author columns.
}
