% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srf_articles.R
\name{separate_srf_authors}
\alias{separate_srf_authors}
\title{A helper function to clean SRF authors}
\usage{
separate_srf_authors(srf_articles)
}
\arguments{
\item{srf_articles}{a data frame containing raw SRF articles as returned from the functions
\code{get_srf_articles} or \code{get_srf_articles_continously}.}
}
\value{
the cleaned data frame \code{srf_articles}
}
\description{
Separates collaborating authors into two separate author columns.
}
