% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woz_articles.R
\name{separate_woz_authors}
\alias{separate_woz_authors}
\title{A helper function to clen WOZ authors}
\usage{
separate_woz_authors(woz_articles)
}
\arguments{
\item{woz_articles}{a data frame containing raw WOZ articles as returned from the functions
\code{get_woz_articles}, \code{get_woz_articles_continously},
or \code{get_woz_articles_from_archive}}
}
\value{
the cleaned data frame \code{woz_articles}
}
\description{
Separates collaborating authors into two separate author columns.
}
